Cinema/CinemaAPI/requirements.txt
flask
gunicorn
flask-restful
google-cloud-firestore
flask-wtf